refactor: 채팅 소켓 관심사 분리
